refactor: fold Xcode into iOS Dev, add asc CLI, remove legacy tools
- Xcode install (via xcodes) now part of iOS Dev module
- Added App Store Connect CLI (brew install asc)
- Removed cocoapods, carthage, fastlane (legacy, SPM replaces them)
- Removed standalone Xcode module
- iOS Dev items: xcode, swiftlint, swiftformat, xcbeautify, asc
